requirements.txt
bcrypt<=4.0.1
cryptography<=41.0.4
fastapi<=0.104.0
pymongo<=4.5.0
python-jose<=3.3.0
uvicorn<=0.23.2
python-dotenv<=1.0.0
passlib<=1.7.4